requirements.txt
python-dotenv
streamlit
pypdf
langchain-community
faiss-cpu
langchain-openai